intializing with the config files now!
