+++ OffSeasonBot2019/src/main/java/com/team2813/frc2019/config/motor/Inverted.java
@@ -1,5 +1,5 @@
 package com.team2813.frc2019.config.motor;
 
 public enum Inverted {
-    INVERTED, FOLLOW_LEADER
+    NONINVERTED, INVERTED, FOLLOW_LEADER
 }
