+++ OffSeasonBot2019/src/main/java/com/team2813/frc2019/config/motor/PeriodicFrame.java
@@ -1,0 +1,19 @@
+package com.team2813.frc2019.config.motor;
+
+import com.revrobotics.CANSparkMaxLowLevel;
+
+public enum PeriodicFrame {
+    STATUS_0, STATUS_1, STATUS_2;
+
+    public CANSparkMaxLowLevel.PeriodicFrame getValue() {
+        switch(this) {
+            case STATUS_0:
+                return CANSparkMaxLowLevel.PeriodicFrame.kStatus0;
+            case STATUS_1:
+                return CANSparkMaxLowLevel.PeriodicFrame.kStatus1;
+            case STATUS_2:
+                return CANSparkMaxLowLevel.PeriodicFrame.kStatus2;
+        }
+        return null;
+    }
+}
